refactor!: move `Policy` trait to jwk module
